requirements.txt
streamlit
torch
torchvision
pillow
numpy
torch
streamlit
